+++ ohsome_quality_api/indicators/currentness/query.sql
@@ -1,15 +1,15 @@
 WITH bpoly AS (
     SELECT
-        ST_Setsrid (ST_GeomFromGeoJSON ('$aoi'), 4326) AS geom
+        ST_SetSRID(ST_GeomFromGeoJSON($$1), 4326) AS geom
 )
 SELECT
-    Date_trunc('month', valid_from) AS month,
-    Count(*)
+    DATE_TRUNC('month', c.valid_from) AS month,
+    $aggregation
 FROM
     $contributions_table c,
     bpoly b
 WHERE
-    ST_Intersects (c.geom, b.geom)
+    ST_Intersects(c.geom, b.geom)
     AND (status_geom_type).status = 'latest'  -- excludes deleted
     AND ($filter)
 GROUP BY